DETECT DELIVER: Fixed orientation when aligning to target.
pep8 format for detect_deliver

un-underscore detect_deliver

DETECT DELIVER: Fixed orientation when aligning to target.
